+++ sdk/js/actra/src/browser-entry.ts
@@ -1,0 +1,32 @@
+//entry for edge & browsers
+
+import { initWasm } from "./init/wasm-init"
+
+// export node actra
+export { ActraCore as Actra } from "./core/actra-core"
+
+// SDK exports
+export { Policy } from "./common/policy"
+export {ActraRuntime} from "./common/runtime"
+export {ActraError, ActraPolicyError} from "./common/errors"
+
+export type {
+  Action,
+  Actor,
+  Snapshot,
+  Decision,
+  DecisionEvent,
+  EvaluationInput,
+  JSONValue
+} from "./common/types"
+
+export type { DecisionObserver } from "./common/events"
+export { setWasmSource } from "./config" //used for wasm source if external CDN, Streaming, ArrayBuffer etc
+export type { ActraWasmSource } from "./config"
+
+import { setWasmLoader } from "./loader/loader-registry"
+import { loadActraWasm } from "./loader/browser-loader"
+
+setWasmLoader(loadActraWasm)
+
+initWasm()